+++ library/optimization/optimizers/__init__.py
@@ -1,24 +1,25 @@
 from library.optimization.optimizers.alice import Alice
 from library.optimization.optimizers.adai import Adai
-from library.optimization.optimizers.adopt import ADOPT, ADOPTMARS, FADOPTMARS
-from library.optimization.optimizers.adopt_schedulefree import (
+from library.optimization.optimizers.adopt import (
+    ADOPT,
     ADOPTEMAMixScheduleFree,
+    ADOPTMARS,
     ADOPTMARSScheduleFree,
     ADOPTNesterovScheduleFree,
+    ADOPTAOScheduleFree,
     ADOPTScheduleFree,
     FADOPTEMAMixScheduleFree,
+    FADOPTMARS,
     FADOPTMARSScheduleFree,
     FADOPTNesterovScheduleFree,
     FADOPTScheduleFree,
 )
-from library.optimization.optimizers.adopt_schedulefree_ao import ADOPTAOScheduleFree
 from library.optimization.optimizers.ademamix import AdEMAMix, SimplifiedAdEMAMix, SimplifiedAdEMAMixExM
 from library.optimization.optimizers.adabelief import AdaBelief
 from library.optimization.optimizers.adan import Adan
-from library.optimization.optimizers.adamw_8bit_kahan import AdamW8bitKahan
-from library.optimization.optimizers.adamw_low_bit import AdamW4bitAO, AdamW8bitAO, AdamWfp8AO
+from library.optimization.optimizers.adamw import AdamW4bitAO, AdamW8bitAO, AdamW8bitKahan, AdamWfp8AO
 from library.optimization.optimizers.adammini import AdamMini
-from library.optimization.optimizers.abmog import ABMOG
+from library.optimization.optimizers.experimental.abmog import ABMOG
 from library.optimization.optimizers.bcos import BCOS
 from library.optimization.optimizers.came import CAME
 from library.optimization.optimizers.cstableadamw import CStableAdamW
@@ -29,47 +30,44 @@ from library.optimization.optimizers.compass import (
     CompassADOPTMARS,
     CompassAO,
     CompassPlus,
+    FCompass,
+    FCompassADOPT,
+    FCompassADOPTMARS,
+    FCompassPlus,
 )
 from library.optimization.optimizers.dehaze import Dehaze
 from library.optimization.optimizers.fftdescent import FFTDescent
-from library.optimization.optimizers.fcompass import FCompass, FCompassADOPT, FCompassADOPTMARS, FCompassPlus
-from library.optimization.optimizers.farmscrop import FARMSCrop
-from library.optimization.optimizers.farmscrop_v2 import FARMSCropV2
-from library.optimization.optimizers.fishmonger import FishMonger, FishMonger8BitBNB
+from library.optimization.optimizers.farmscrop import FARMSCrop, FARMSCropV2
+from library.optimization.optimizers.experimental.fishmonger import FishMonger, FishMonger8BitBNB
 from library.optimization.optimizers.fira import Fira
-from library.optimization.optimizers.fmarscrop import FMARSCrop
-from library.optimization.optimizers.fmarscrop_v2 import FMARSCropV2
-from library.optimization.optimizers.fmarscrop_v2_exmachina import FMARSCropV2ExMachina
-from library.optimization.optimizers.fmarscrop_v3 import FMARSCropV3
-from library.optimization.optimizers.fmarscrop_v3_exmachina import FMARSCropV3ExMachina
+from library.optimization.optimizers.fmarscrop import FMARSCrop, FMARSCropV2, FMARSCropV2ExMachina, FMARSCropV3, FMARSCropV3ExMachina
 from library.optimization.optimizers.galore import GaLore
 from library.optimization.optimizers.glyph import Glyph
-from library.optimization.optimizers.gooddog import GOODDOG
+from library.optimization.optimizers.experimental.gooddog import GOODDOG
 from library.optimization.optimizers.grokfast import GrokFastAdamW
 from library.optimization.optimizers.laprop import LaProp
 from library.optimization.optimizers.lamb import Lamb
 from library.optimization.optimizers.lpf_adamw import LPFAdamW
-from library.optimization.optimizers.mythical import Mythical
-from library.optimization.optimizers.momentus_caution import MomentusCaution
-from library.optimization.optimizers.oagopt import OAGOpt
-from library.optimization.optimizers.ocgopt import OCGOpt
+from library.optimization.optimizers.experimental.mythical import Mythical
+from library.optimization.optimizers.experimental.momentus_caution import MomentusCaution
+from library.optimization.optimizers.experimental.oagopt import OAGOpt
+from library.optimization.optimizers.experimental.ocgopt import OCGOpt
 from library.optimization.optimizers.projective_adam import ProjectiveAdam
 from library.optimization.optimizers.racs import RACS
 from library.optimization.optimizers.ranger21 import Ranger21
 from library.optimization.optimizers.rmsprop import RMSProp, RMSPropADOPT, RMSPropADOPTMARS
-from library.optimization.optimizers.scgopt import SCGOpt
+from library.optimization.optimizers.experimental.scgopt import SCGOpt
 from library.optimization.optimizers.scion import SCION
-from library.optimization.optimizers.scorn import SCORN
-from library.optimization.optimizers.scornmachina import SCORNMachina
+from library.optimization.optimizers.scorn import SCORN, SCORNMachina
 from library.optimization.optimizers.sgd_sai import SGDSaI
 from library.optimization.optimizers.shampoo import ScalableShampoo
-from library.optimization.optimizers.singstate import SingState
+from library.optimization.optimizers.experimental.singstate import SingState
 from library.optimization.optimizers.soap import SOAP
 from library.optimization.optimizers.spam import StableSPAM
-from library.optimization.optimizers.remaster import REMASTER
-from library.optimization.optimizers.talon import TALON
+from library.optimization.optimizers.experimental.remaster import REMASTER
+from library.optimization.optimizers.experimental.talon import TALON
 from library.optimization.optimizers.vsgd import VSGD
-from library.optimization.optimizers.wiwiopt import WiwiOpt
+from library.optimization.optimizers.experimental.wiwiopt import WiwiOpt
 
 
 __all__ = [