+++ library/optimization/optimizers/compass/__init__.py
@@ -1,0 +1,23 @@
+from library.optimization.optimizers.compass.compass import (
+    Compass,
+    Compass8BitBNB,
+    CompassADOPT,
+    CompassADOPTMARS,
+    CompassAO,
+    CompassPlus,
+)
+from library.optimization.optimizers.compass.fcompass import FCompass, FCompassADOPT, FCompassADOPTMARS, FCompassPlus
+
+
+__all__ = [
+    "Compass",
+    "Compass8BitBNB",
+    "CompassADOPT",
+    "CompassADOPTMARS",
+    "CompassAO",
+    "CompassPlus",
+    "FCompass",
+    "FCompassADOPT",
+    "FCompassADOPTMARS",
+    "FCompassPlus",
+]