+++ library/optimization/optimizers/experimental/gooddog.py
@@ -1,3 +1,5 @@
+# GOODDOG from https://github.com/Clybius/Personalized-Optimizers by Clybius
+
 import torch
 from torch.optim import Optimizer